Make puppeteer-parse a module
