--- packages/puppeteer-parse/.env.example
@@ -1,8 +1,0 @@
-# Should match with the JWT_SECRET that the api uses
-JWT_SECRET=some_secret
-
-# Address of the backend that is running locally
-REST_BACKEND_ENDPOINT=http://localhost:4000/api
-
-# set for local development
-IS_LOCAL=true
