--- packages/puppeteer-parse/.gitignore
@@ -1,1 +1,0 @@
-previewImage.*
