+++ packages/puppeteer-parse/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@omnivore/puppeteer-parse",
   "version": "1.0.0",
-  "description": "Google Cloud Function that accepts URL of the article and parses its content",
+  "description": "Accepts URL of the article and parses its content",
   "main": "index.js",
   "dependencies": {
     "@google-cloud/logging-winston": "^4.1.2",
@@ -22,8 +22,6 @@
     "@google-cloud/functions-framework": "^3.0.0"
   },
   "scripts": {
-    "start": "npx functions-framework --port=9090 --target=puppeteer",
-    "start_preview": "npx functions-framework --target=preview",
     "test": "yarn mocha"
   }
 }
